refactor cartItem request

+++ PharmaShop-server/pharmacy.Application/Services/OrderService.cs
@@ -1,10 +1,9 @@
 ﻿using AutoMapper;
 using pharmacy.Core.Contracts.IServices;
-using pharmacy.Core.Contracts;
 using pharmacy.Core.DTOs.Order;
-using pharmacy.Core.Entities;
 using pharmacy.Core;
-namespace pharmacy.Application.Services;
+using pharmacy.Core.Entities;
+
 public class OrderService : IOrderService
 {
     private readonly IUnitOfWork _unitOfWork;
@@ -15,11 +14,25 @@ public class OrderService : IOrderService
         _unitOfWork = unitOfWork;
         _mapper = mapper;
     }
+    public async Task<OrderResponseDto> CreateAsync(OrderRequestDto request)
+    {
+        var order = _mapper.Map<Order>(request);
+
+        await _unitOfWork.orderRepository.CreateAsync(order);
+
+        _unitOfWork.Complete();
+
+        return _mapper.Map<OrderResponseDto>(order);
+    }
+
 
     public async Task<OrderResponseDto> GetByIdAsync(int id)
     {
         var order = await _unitOfWork.orderRepository.GetByID(id);
-        _unitOfWork.Complete();
+        if (order == null)
+        {
+            throw new Exception("Order not found");
+        }
         return _mapper.Map<OrderResponseDto>(order);
     }
 
@@ -29,33 +42,30 @@ public class OrderService : IOrderService
         return _mapper.Map<IEnumerable<OrderResponseDto>>(orders);
     }
 
-    public async Task<OrderResponseDto> CreateAsync(OrderRequestDto request)
-    {
-        var orderEntity = _mapper.Map<Order>(request);
-        orderEntity.CreatedDate = DateTime.UtcNow;
-
-        var createdOrder = await _unitOfWork.orderRepository.CreateAsync(orderEntity);
-        _unitOfWork.Complete();
-        return _mapper.Map<OrderResponseDto>(createdOrder);
-    }
-
 
     public async Task<OrderResponseDto> UpdateAsync(int id, OrderRequestDto request)
     {
         var existingOrder = await _unitOfWork.orderRepository.GetByID(id);
-        if (existingOrder is null)
+        if (existingOrder == null)
+        {
             throw new Exception("Order not found");
+        }
 
         _mapper.Map(request, existingOrder);
-        await _unitOfWork.orderRepository.UpdateAsync(id,existingOrder);
+        await _unitOfWork.orderRepository.UpdateAsync(id, existingOrder);
         _unitOfWork.Complete();
 
         return _mapper.Map<OrderResponseDto>(existingOrder);
     }
 
-
     public async Task<string> DeleteAsync(int id)
     {
+        var order = await _unitOfWork.orderRepository.GetByID(id);
+        if (order == null)
+        {
+            throw new Exception("Order not found");
+        }
+
         await _unitOfWork.orderRepository.DeleteAsync(id);
         _unitOfWork.Complete();
         return "Deleted success";
